+++ packages/stateless/hooks/index.ts
@@ -9,6 +9,7 @@ export * from './useDaoNavHelpers'
 export * from './useDetectTruncate'
 export * from './useDetectWrap'
 export * from './useGovProposalVoteOptions'
+export * from './useInfiniteScroll'
 export * from './useLoadingGovProposalTimestampInfo'
 export * from './useMountedInBrowser'
 export * from './usePlatform'
